Let channel control message form and values.
